Add new shader source to allow creating shaders from binary streamas
